Move the turn_search derivation SQL into llm/logs.py
TURN_SEARCH_LITERAL and TURN_SEARCH_INSERT_SQL lived in migrations.py
for the migration backfill, which no longer exists - the per-turn
refresh in LogStore.log is their only consumer, so they move next to
it. The turn_filter format slot goes too: with one caller left, the
:turn_id filter is baked into the SQL directly.
